+++ src/packages/media/media-types/repository/detail/index.ts
@@ -1,3 +1,3 @@
 export { UmbMediaTypeDetailRepository } from './media-type-detail.repository.js';
 export { UMB_MEDIA_TYPE_DETAIL_REPOSITORY_ALIAS, UMB_MEDIA_TYPE_DETAIL_STORE_ALIAS } from './constants.js';
-export { UMB_MEDIA_TYPE_DETAIL_STORE_CONTEXT } from './media-type-detail.store.js';
+export { UMB_MEDIA_TYPE_DETAIL_STORE_CONTEXT } from './media-type-detail.store.context-token.js';
