+++ src/packages/media/media-types/repository/detail/media-type-detail.store.context-token.ts
@@ -1,0 +1,6 @@
+import type UmbMediaTypeDetailStore from './media-type-detail.store.js';
+import { UmbContextToken } from '@umbraco-cms/backoffice/context-api';
+
+export const UMB_MEDIA_TYPE_DETAIL_STORE_CONTEXT = new UmbContextToken<UmbMediaTypeDetailStore>(
+	'UmbMediaTypeDetailStore',
+);
