+++ src/packages/media/media-types/repository/item/index.ts
@@ -1,4 +1,4 @@
 export { UmbMediaTypeItemRepository } from './media-type-item.repository.js';
 export { UMB_MEDIA_TYPE_ITEM_REPOSITORY_ALIAS, UMB_MEDIA_TYPE_ITEM_STORE_ALIAS } from './constants.js';
-export { UMB_MEDIA_TYPE_ITEM_STORE_CONTEXT } from './media-type-item.store.js';
+export { UMB_MEDIA_TYPE_ITEM_STORE_CONTEXT } from './media-type-item.store.context-token.js';
 export type { UmbMediaTypeItemModel } from './types.js';
