+++ src/packages/media/media-types/repository/item/media-type-item.store.context-token.ts
@@ -1,0 +1,4 @@
+import type UmbMediaTypeItemStore from './media-type-item.store.js';
+import { UmbContextToken } from '@umbraco-cms/backoffice/context-api';
+
+export const UMB_MEDIA_TYPE_ITEM_STORE_CONTEXT = new UmbContextToken<UmbMediaTypeItemStore>('UmbMediaTypeItemStore');
