+++ src/packages/media/media-types/tree/index.ts
@@ -5,7 +5,7 @@ export {
 } from './constants.js';
 export * from './folder/index.js';
 export { UmbMediaTypeTreeRepository } from './media-type-tree.repository.js';
-export { UMB_MEDIA_TYPE_TREE_STORE_CONTEXT } from './media-type-tree.store.js';
+export { UMB_MEDIA_TYPE_TREE_STORE_CONTEXT } from './media-type-tree.store.context-token.js';
 export { UMB_MEDIA_TYPE_PICKER_MODAL } from './media-type-picker-modal.token.js';
 
 export type { UmbMediaTypeTreeItemModel, UmbMediaTypeTreeRootModel } from './types.js';
