+++ src/packages/media/media-types/tree/media-type-tree.store.context-token.ts
@@ -1,0 +1,4 @@
+import type UmbMediaTypeTreeStore from './media-type-tree.store.js';
+import { UmbContextToken } from '@umbraco-cms/backoffice/context-api';
+
+export const UMB_MEDIA_TYPE_TREE_STORE_CONTEXT = new UmbContextToken<UmbMediaTypeTreeStore>('UmbMediaTypeTreeStore');
