+++ src/packages/media/media-types/tree/media-type-tree.store.ts
@@ -1,4 +1,4 @@
-import { UmbContextToken } from '@umbraco-cms/backoffice/context-api';
+import { UMB_MEDIA_TYPE_TREE_STORE_CONTEXT } from './media-type-tree.store.context-token.js';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
 import { UmbUniqueTreeStore } from '@umbraco-cms/backoffice/tree';
 
@@ -20,5 +20,3 @@ export class UmbMediaTypeTreeStore extends UmbUniqueTreeStore {
 }
 
 export default UmbMediaTypeTreeStore;
-
-export const UMB_MEDIA_TYPE_TREE_STORE_CONTEXT = new UmbContextToken<UmbMediaTypeTreeStore>('UmbMediaTypeTreeStore');
