+++ src/packages/media/media/repository/item/media-item.repository.ts
@@ -1,5 +1,5 @@
 import { UmbMediaItemServerDataSource } from './media-item.server.data-source.js';
-import { UMB_MEDIA_ITEM_STORE_CONTEXT } from './media-item.store.js';
+import { UMB_MEDIA_ITEM_STORE_CONTEXT } from './media-item.store.context-token.js';
 import type { UmbMediaItemModel } from './types.js';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
 import { UmbItemRepositoryBase } from '@umbraco-cms/backoffice/repository';
