+++ src/packages/media/media/repository/item/media-item.store.context-token.ts
@@ -1,0 +1,4 @@
+import type UmbMediaItemStore from './media-item.store.js';
+import { UmbContextToken } from '@umbraco-cms/backoffice/context-api';
+
+export const UMB_MEDIA_ITEM_STORE_CONTEXT = new UmbContextToken<UmbMediaItemStore>('UmbMediaItemStore');
