+++ src/packages/media/media/repository/item/media-item.store.ts
@@ -1,5 +1,5 @@
 import type { UmbMediaDetailModel } from '../../types.js';
-import { UmbContextToken } from '@umbraco-cms/backoffice/context-api';
+import { UMB_MEDIA_ITEM_STORE_CONTEXT } from './media-item.store.context-token.js';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
 import { UmbItemStoreBase } from '@umbraco-cms/backoffice/store';
 
@@ -22,5 +22,3 @@ export class UmbMediaItemStore extends UmbItemStoreBase<UmbMediaDetailModel> {
 }
 
 export default UmbMediaItemStore;
-
-export const UMB_MEDIA_ITEM_STORE_CONTEXT = new UmbContextToken<UmbMediaItemStore>('UmbMediaItemStore');
