+++ src/packages/media/media/repository/url/media-url.repository.ts
@@ -1,5 +1,5 @@
 import type { UmbMediaUrlModel } from './types.js';
-import { UMB_MEDIA_URL_STORE_CONTEXT } from './media-url.store.js';
+import { UMB_MEDIA_URL_STORE_CONTEXT } from './media-url.store.context-token.js';
 import { UmbMediaUrlServerDataSource } from './media-url.server.data-source.js';
 import { UmbItemRepositoryBase } from '@umbraco-cms/backoffice/repository';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
