+++ src/packages/media/media/repository/url/media-url.store.ts
@@ -1,5 +1,5 @@
 import type { UmbMediaDetailModel } from '../../types.js';
-import { UmbContextToken } from '@umbraco-cms/backoffice/context-api';
+import { UMB_MEDIA_URL_STORE_CONTEXT } from './media-url.store.context-token.js';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
 import { UmbItemStoreBase } from '@umbraco-cms/backoffice/store';
 
@@ -22,5 +22,3 @@ export class UmbMediaUrlStore extends UmbItemStoreBase<UmbMediaDetailModel> {
 }
 
 export default UmbMediaUrlStore;
-
-export const UMB_MEDIA_URL_STORE_CONTEXT = new UmbContextToken<UmbMediaUrlStore>('UmbMediaUrlStore');
