+++ src/packages/media/media/tree/media-tree.store.context-token.ts
@@ -1,0 +1,4 @@
+import type UmbMediaTreeStore from './media-tree.store.js';
+import { UmbContextToken } from '@umbraco-cms/backoffice/context-api';
+
+export const UMB_MEDIA_TREE_STORE_CONTEXT = new UmbContextToken<UmbMediaTreeStore>('UmbMediaTreeStore');
