+++ src/packages/media/media/tree/media-tree.store.ts
@@ -1,4 +1,4 @@
-import { UmbContextToken } from '@umbraco-cms/backoffice/context-api';
+import { UMB_MEDIA_TREE_STORE_CONTEXT } from './media-tree.store.context-token.js';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
 import { UmbUniqueTreeStore } from '@umbraco-cms/backoffice/tree';
 
@@ -20,5 +20,3 @@ export class UmbMediaTreeStore extends UmbUniqueTreeStore {
 }
 
 export default UmbMediaTreeStore;
-
-export const UMB_MEDIA_TREE_STORE_CONTEXT = new UmbContextToken<UmbMediaTreeStore>('UmbMediaTreeStore');
